+++ src/main/java/com/ecommerce/project/payload/OrderItemDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,6 +13,6 @@ public class OrderItemDTO {
     private Long orderItemId;
     private ProductDTO product;
     private Integer quantity;
-    private Double discount;
-    private Double orderedProductPrice;
+    private BigDecimal discount;
+    private BigDecimal orderedProductPrice;
 }
